backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==11.0.3
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pydantic==2.4.2
python-dotenv==1.0.0
transformers==4.37.2
kokoro==0.7.16
spaces
psutil==5.9.8
